Move dumper to utils.py
